refactor: Move BooleanFlag to the only place it's used
